requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.0.0
jsonschema>=4.20.0
python-multipart>=0.0.6
openai>=1.0.0
anthropic>=0.7.0
requests>=2.31.0
python-dotenv>=1.0.0

